# apps/backend/openapi.json
{
  "openapi": "3.1.1",
  "paths": {
...
    "/users": {
      "get": {
        "description": "Only contains a simplified view of a user",
        "operationId": "UserController_getUsers",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "role",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Users successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SimpleUserDto" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        },
        "security": [{ "apiKeyCookie": [] }],
        "summary": "List all users (admin only)",
        "tags": ["Users"]
      }
    },
    "/users/{userId}": {
      "get": {
        "operationId": "UserController_getUserById",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully retrieved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserDto" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "User not found" }
        },
        "security": [{ "apiKeyCookie": [] }],
        "summary": "Get detailed user information (admin only)",
        "tags": ["Users"]
      }
    },
    "/jobs": {
      "get": {
        "operationId": "JobController_getJobs",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/JobDto" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        },
        "security": [{ "apiKeyCookie": [] }],
        "summary": "List jobs (filtered by user unless admin)",
        "tags": ["Jobs"]
      }
    },
    "/jobs/{jobId}": {
      "get": {
        "operationId": "JobController_getJobById",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Job successfully retrieved.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobDto" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Job not found" }
        },
        "security": [{ "apiKeyCookie": [] }],
        "summary": "Get detailed job information",
        "tags": ["Jobs"]
      }
    },
    "/jobs/{jobId}/result": {
      "get": {
        "operationId": "JobController_getJobResult",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Job result successfully retrieved.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobResultDto" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Job or result not found" }
        },
        "security": [{ "apiKeyCookie": [] }],
        "summary": "Get job result output",
        "tags": ["Jobs"]
      }
    },
    "/emails": {
      "post": {
        "operationId": "EmailController_submitEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateEmailDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully submitted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "data": { "type": "string" } },
                  "required": ["data"]
                }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "500": {
            "description": "AiBackend processing failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "details": { "type": "string" }
                  }
                }
              }
            }
          }
        },
        "security": [{ "apiKeyCookie": [] }],
        "summary": "Submit an email for processing",
        "tags": ["Emails"]
      }
    }
  },
  "info": {
    "title": "SIEVE",
    "description": "This is the OpenAPI specification for the communication to the SIEVE backend via the REST-API. The SIEVE frontend and backend communicate via this API.\n\n**Note**: Better Auth currently does not support customizing the tag names.\nTherefore, the 'Default' tag is used for endpoints that are related to user authentication and session management.",
    "version": "0.1.0",
    "contact": {},
    "license": {
      "name": "GPLv3-or-later",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    }
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SimpleUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user ID",
            "example": "ff1fa1cb-d247-43ec-87d7-4033ae527b2d"
          },
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Email address of the user",
            "example": "john.doe@example.com"
          },
          "role": {
            "type": "string",
            "description": "Role of the user",
            "enum": ["ADMIN", "USER"],
            "example": "USER"
          }
        },
        "required": ["id", "name", "email", "role"]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user ID",
            "example": "ff1fa1cb-d247-43ec-87d7-4033ae527b2d"
          },
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Email address of the user",
            "example": "john.doe@example.com"
          },
          "role": {
            "type": "string",
            "description": "Role of the user",
            "enum": ["ADMIN", "USER"],
            "example": "USER"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last update timestamp",
            "format": "date-time",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "description": "Deletion timestamp, if soft-deleted",
            "format": "date-time",
            "nullable": true
          },
          "jobs": {
            "description": "IDs of jobs created by the user",
            "type": "array",
            "items": { "type": "string" }
          },
          "emails": {
            "description": "IDs of emails uploaded by the user",
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["id", "name", "email", "role", "createdAt"]
      },
      "JobDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique job ID",
            "example": "ff1fa1cb-d247-43ec-87d7-4033ae527b2d"
          },
          "status": {
            "type": "string",
            "description": "Current status of the job",
            "enum": ["CREATED", "PENDING", "PROCESSING", "COMPLETED"],
            "example": "CREATED"
          },
          "userId": {
            "type": "string",
            "description": "User ID who created the job",
            "example": "ff1fa1cb-d247-43ec-87d7-4033ae527b2d"
          },
          "emailId": {
            "type": "string",
            "description": "Email ID associated with this job",
            "example": "ff1fa1cb-d247-43ec-87d7-4033ae527b2d"
          },
          "startedAt": {
            "type": "string",
            "description": "Job start timestamp",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "description": "Job completion timestamp",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "description": "Deletion timestamp if soft-deleted",
            "format": "date-time",
            "nullable": true
          },
          "resultId": {
            "type": "string",
            "description": "Associated JobResult ID, if completed",
            "nullable": true
          }
        },
        "required": ["id", "status", "userId", "emailId", "createdAt"]
      },
      "JobResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique JobResult ID",
            "example": "ff1fa1cb-d247-43ec-87d7-4033ae527b2d"
          },
          "jobId": {
            "type": "string",
            "description": "Associated Job ID",
            "example": "ff1fa1cb-d247-43ec-87d7-4033ae527b2d"
          },
          "status": {
            "type": "string",
            "description": "Status of the job result",
            "enum": ["SUCCESS", "FAILURE"],
            "example": "SUCCESS"
          },
          "output": {
            "type": "object",
            "description": "JSON output of the job"
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "description": "Deletion timestamp if soft-deleted",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": ["id", "jobId", "status", "output", "createdAt"]
      },
      "CreateEmailDto": {
        "type": "object",
        "properties": {
          "sender": {
            "type": "string",
            "description": "Sender of the email",
            "example": "alice.smith@example.com",
            "nullable": true
          },
          "subject": {
            "type": "string",
            "description": "Subject of the email",
            "example": "Piece Availability Request",
            "nullable": true
          },
          "body": {
            "type": "string",
            "description": "Body of the email",
            "example": "Dear Sir or Madam,\n\nI would like to request information regarding the availability of the following pieces...\n\nBest regards,\nAlice Smith"
          }
        },
        "required": ["body"]
      }
    },
    "securitySchemes": {
      "apiKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "apiKeyCookie",
        "description": "API Key authentication via cookie"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token authentication"
      }
    }
  },
...
  "security": [{ "apiKeyCookie": [], "bearerAuth": [] }]
}